+++ src/main/scala/vectorpipe/functions/osm/package.scala
@@ -1,14 +1,13 @@
 package vectorpipe.functions
 
 import org.apache.spark.sql.expressions.UserDefinedFunction
-import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
-import org.apache.spark.sql.{Column, Row}
+import org.apache.spark.sql.{Column, DataFrame, Row}
 import vectorpipe.model.Member
 
-import scala.util.{Try, Success, Failure}
 import scala.util.matching.Regex
+import scala.util.{Failure, Success, Try}
 
 package object osm {
   // Using tag listings from [id-area-keys](https://github.com/osmlab/id-area-keys) @ v2.13.0.
@@ -142,46 +141,63 @@ package object osm {
 
   private val MultiPolygonTypes = Seq("multipolygon", "boundary")
 
-  private val BooleanValues = Seq("yes", "no", "true", "false", "1", "0")
-
   private val TruthyValues = Seq("yes", "true", "1")
+
+  private val FalsyValues = Seq("no", "false", "0")
+
+  private val BooleanValues = TruthyValues ++ FalsyValues
 
   private val WaterwayValues =
     Seq(
-      "river", "canal", "stream", "brook", "drain", "ditch"
+      "river", "riverbank", "canal", "stream", "stream_end", "brook", "drain", "ditch", "dam", "weir", "waterfall",
+      "pressurised"
     )
 
   private val POITags = Set("amenity", "shop", "craft", "office", "leisure", "aeroway")
 
   private val HashtagMatcher: Regex = """#([^\u2000-\u206F\u2E00-\u2E7F\s\\'!\"#$%()*,.\/;<=>?@\[\]^{|}~]+)""".r
 
+  private def cleanDelimitedValues(values: Column): Column = regexp_replace(trim(values), "\\s*;\\s*", ";")
+
+  def splitDelimitedValues(values: Column, default: Column = lit("")): Column = split(lower(coalesce(cleanDelimitedValues(values), default)), ";")
+
+  def splitDelimitedValues(values: String): Set[String] = values.replaceAll("\\s*;\\s*", ";").toLowerCase().split(";").toSet
+
   private val _isArea = (tags: Map[String, String]) =>
     tags match {
-      case _ if tags.contains("area") && BooleanValues.contains(tags("area").toLowerCase) =>
-        TruthyValues.contains(tags("area").toLowerCase)
+      case _ if tags.contains("area") && BooleanValues.toSet.intersect(splitDelimitedValues(tags("area"))).nonEmpty =>
+
+        TruthyValues.toSet.intersect(splitDelimitedValues(tags("area"))).nonEmpty
       case _ =>
         // see https://github.com/osmlab/id-area-keys (values are inverted)
         val matchingKeys = tags.keySet.intersect(AreaKeys.keySet)
-        matchingKeys.exists(k => !AreaKeys(k).contains(tags(k)))
+
+        matchingKeys.exists(k => {
+          // break out semicolon-delimited values
+          val values = splitDelimitedValues(tags(k))
+
+          // values that should be considered as lines
+          AreaKeys(k).keySet
+            .intersect(values)
+            // at least one key passes the area test
+            .size < values.size
+        })
     }
 
-  @transient lazy val isAreaUDF: UserDefinedFunction = udf(_isArea)
+  val isAreaUDF: UserDefinedFunction = udf(_isArea)
 
-  def isArea(tags: Column): Column =
-    when(lower(coalesce(tags.getField("area"), lit(""))).isin(BooleanValues: _*),
-         lower(tags.getField("area")).isin(TruthyValues: _*))
-      // only call the UDF when necessary
-      .otherwise(isAreaUDF(tags)) as 'isArea
+  def isArea(tags: Column): Column = isAreaUDF(tags) as 'isArea
 
   def isMultiPolygon(tags: Column): Column =
-    lower(coalesce(tags.getItem("type"), lit("")))
-      .isin(MultiPolygonTypes: _*) as 'isMultiPolygon
+    array_intersects(
+      splitDelimitedValues(tags.getItem("type")),
+      lit(MultiPolygonTypes.toArray)) as 'isMultiPolygon
 
   def isNew(version: Column, minorVersion: Column): Column =
     version <=> 1 && minorVersion <=> 0 as 'isNew
 
   def isRoute(tags: Column): Column =
-    lower(tags.getItem("type")) <=> "route" as 'isRoute
+    array_contains(splitDelimitedValues(tags.getItem("type")), "route") as 'isRoute
 
   private lazy val MemberSchema = ArrayType(
     StructType(
@@ -261,7 +277,7 @@ package object osm {
       .otherwise(typedLit(Seq.empty[String])) as 'hashtags
 
   def isBuilding(tags: Column): Column =
-    lower(coalesce(tags.getItem("building"), lit("no"))) =!= "no" as 'isBuilding
+    !lower(coalesce(tags.getItem("building"), lit("no"))).isin(FalsyValues: _*) as 'isBuilding
 
   @transient lazy val isPOI: UserDefinedFunction = udf {
     tags: Map[String, String] => POITags.intersect(tags.keySet).nonEmpty
@@ -271,13 +287,15 @@ package object osm {
     tags.getItem("highway").isNotNull as 'isRoad
 
   def isCoastline(tags: Column): Column =
-    lower(tags.getItem("natural")) <=> "coastline" as 'isCoastline
+    array_contains(splitDelimitedValues(tags.getItem("natural")), "coastline") as 'isCoastline
 
   def isWaterway(tags: Column): Column =
-    lower(coalesce(tags.getItem("waterway"), lit("")))
-      .isin(WaterwayValues: _*) as 'isWaterway
+    array_intersects(splitDelimitedValues(tags.getItem("waterway")), lit(WaterwayValues.toArray)) as 'isWaterway
 
   def mergeTags: UserDefinedFunction = udf {
     (_: Map[String, String]) ++ (_: Map[String, String])
   }
+
+  val array_intersects: UserDefinedFunction = udf { (a: Seq[_], b: Seq[_]) =>
+    a.intersect(b).nonEmpty}
 }